+++ src/filters/feed_filters.ts
@@ -2,7 +2,6 @@
  * Helpers for building and serializing a complete set of FeedFilterSettings.
  */
 import BooleanFilter, { TYPE_FILTERS, BooleanFilterArgs, isBooleanFilterName } from "./boolean_filter";
-import BooleanFilterOptionList, { HashtagFilterOptionList, LanguageFilterOptionList, UserFilterOptionList } from "./boolean_filter_option_list";
 import NumericFilter, { FILTERABLE_SCORES, isNumericFilterName } from "./numeric_filter";
 import Storage from "../Storage";
 import Toot from "../api/objects/toot";
@@ -10,6 +9,11 @@ import { BooleanFilterName } from '../enums';
 import { config } from "../config";
 import { incrementCount, split, sumArray, sumValues } from "../helpers/collection_helpers";
 import { Logger } from '../helpers/logger';
+import BooleanFilterOptionList, {
+    HashtagFilterOptionList,
+    LanguageFilterOptionList,
+    UserFilterOptionList
+} from "./boolean_filter_option_list";
 import {
     type BooleanFilters,
     type FeedFilterSettings,